requirements.txt
aiohttp==3.8.1
async-timeout==4.0.2
attrs==21.4.0
chardet==5.0.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
typing_extensions==4.3.0
yarl==1.7.2
